ready

+++ src/database_controller/controller_database.py
@@ -7,27 +7,27 @@ from gensim.corpora import Dictionary
 from database_controller.database_interface import Controller
 from searcher.preprocess import data_processing
 
-def read_or_create_joblib(ip):
+def create_joblib(ip):
     ip = str(ip)
     os.makedirs(f"src/data/{ip}/", exist_ok=True)
 
-    if os.path.exists(f"src/data/{ip}/dictionary.joblib"):
-        return load(f"src/data/{ip}/dictionary.joblib")
-    else:
-        dump(Dictionary(), f"src/data/{ip}/dictionary.joblib")
-        return Dictionary()
+    for file in os.listdir(f"src/data/{ip}/"):
+        os.remove(f"src/data/{ip}/{file}")
+
+    dump(Dictionary(), f"src/data/{ip}/dictionary.joblib")
+    return Dictionary()
 
 class DocumentController(Controller):
     dictionary: Dictionary
     
     def __init__(self, ip):
         self.ip = ip
-        DocumentController.dictionary = read_or_create_joblib(ip)
+        DocumentController.dictionary = create_joblib(ip)
 
     def connect(self):
         return sqlite3.connect(f"src/data/{self.ip}/database.db")
 
-    def create_document(self, id, text, table):
+    def create_document(self, id, text, clock, table):
         try:
             if table == 'documentos':
                 tokens = data_processing.tokenize_corpus([text])
@@ -41,8 +41,8 @@ class DocumentController(Controller):
 
 
                 cursor.execute(f'''
-                    INSERT INTO {table} (id, text, tf) VALUES (?, ?, ?)
-                ''', (id, text, tf_json))
+                    INSERT INTO {table} (id, text, clock, tf) VALUES (?, ?, ?, ?)
+                ''', (id, text, clock, tf_json))
                 conn.commit()
                 conn.close()
 
@@ -52,8 +52,8 @@ class DocumentController(Controller):
                 cursor = conn.cursor()
 
                 cursor.execute(f'''
-                    INSERT INTO {table} (id, text) VALUES (?, ?)
-                ''', (id, text))
+                    INSERT INTO {table} (id, text, clock) VALUES (?, ?, ?)
+                ''', (id, text, clock))
                 conn.commit()
                 conn.close()
         except:
@@ -69,14 +69,16 @@ class DocumentController(Controller):
         
         return docs
     
-    def get_document_by_id(self, _id):
-        conn = self.connect()
-        cursor = conn.cursor()
-        cursor.execute('SELECT text FROM documentos WHERE id = ?', (_id,))
-        doc = cursor.fetchone()
-        conn.close()
-        
-        return doc[0]
+    def get_document_by_id(self, _id, table = 'documentos'):
+        try:
+            conn = self.connect()
+            cursor = conn.cursor()
+            cursor.execute(f'SELECT text, clock FROM {table} WHERE id = ?', (_id,))
+            doc = cursor.fetchone()
+            conn.close()
+            return doc[0], doc[1]
+        except:
+            return None
     
     def get_documents_for_query(self):
         conn = self.connect()
@@ -87,7 +89,7 @@ class DocumentController(Controller):
         
         return doc
 
-    def update_document(self, id, table, text=None):
+    def update_document(self, id, clock, table, text=None):
         conn = self.connect()
         cursor = conn.cursor()
 
@@ -105,8 +107,8 @@ class DocumentController(Controller):
                     DocumentController.dictionary.cfs[word] -= count
                     DocumentController.dictionary.dfs[word] -= 1
                 cursor.execute(f'''
-                    UPDATE {table} SET text = ? WHERE id = ?
-                ''', (text, id))
+                    UPDATE {table} SET text = ?, clock = ? WHERE id = ?
+                ''', (text, clock, id))
             
             tokens_text = data_processing.tokenize_corpus([text])
             tf = DocumentController.dictionary.doc2bow(tokens_text[0])
@@ -123,8 +125,8 @@ class DocumentController(Controller):
 
         else:
             cursor.execute(f'''
-                    UPDATE {table} SET text = ? WHERE id = ?
-                ''', (text, id))
+                    UPDATE {table} SET text = ?, clock = ? WHERE id = ?
+                ''', (text, clock, id))
             
         conn.commit()
         conn.close()
@@ -135,18 +137,20 @@ class DocumentController(Controller):
 
         if table == 'documentos':
             cursor.execute(f'SELECT text FROM {table} WHERE id = ?', (id,))
-            
-            doc = cursor.fetchone()[0]
-            tokens = data_processing.tokenize_corpus([doc])
-            
-            bow = DocumentController.dictionary.doc2bow(tokens[0])
-            
-            for word, count in bow:
-                DocumentController.dictionary.cfs[word] -= count
-                DocumentController.dictionary.dfs[word] -= 1
-            
-            cursor = conn.cursor()
-            cursor.execute(f'DELETE FROM {table} WHERE id = ?', (id,))
+
+            doc = cursor.fetchone()
+            if doc:
+                doc = doc[0]
+                tokens = data_processing.tokenize_corpus([doc])
+                
+                bow = DocumentController.dictionary.doc2bow(tokens[0])
+                
+                for word, count in bow:
+                    DocumentController.dictionary.cfs[word] -= count
+                    DocumentController.dictionary.dfs[word] -= 1
+                
+                cursor = conn.cursor()
+                cursor.execute(f'DELETE FROM {table} WHERE id = ?', (id,))
 
         else:
             cursor.execute(f'DELETE FROM {table} WHERE id = ?', (id,))
